+++ src/frame.rs
@@ -1,8 +1,8 @@
 pub struct Frame {
-    pub x: u16,
-    pub y: u16,
-    pub width: u16,
-    pub height: u16,
+    pub x: i32,
+    pub y: i32,
+    pub width: i32,
+    pub height: i32,
 }
 
 impl Frame {
@@ -11,8 +11,8 @@ impl Frame {
         Frame {
             x: 0,
             y: 0,
-            width: width,
-            height: height,
+            width: width as i32,
+            height: height as i32,
         }
     }
 }
